Add extensions for spilling solutions into puddles
