+++ Content.Server/GameObjects/Components/Fluids/CanSpillComponent.cs
@@ -41,7 +41,7 @@ namespace Content.Server.GameObjects.Components.Fluids
                 // Need this as when we split the component's owner may be deleted
                 var entityLocation = component.Owner.Transform.GridPosition;
                 var solution = solutionComponent.SplitSolution(solutionComponent.CurrentVolume);
-                SpillHelper.SpillAt(entityLocation, solution, "PuddleSmear");
+                solution.SpillAt(entityLocation, "PuddleSmear");
             }
         }
     }
